Add files via upload
This repository contains assignments done.

Assignment 1: A raw and messy data was provided on a marketing firm, which I then performed data wrangling in order to clean it up. The cleaned/sorted data was then written to a csv file. I wrangled further and created a csv file .The most and least performing agents, month with the highest sales are also contained in the file.

Assignmennt 2: This assignment was based on mining data off a web page. I was tasked with mining certain information of all smartphones on jumia(an e-commerce platform) and writing the data to a csv file.

Asssignment 3: This was based on the previous assignment. I wrote the data mined to a relational database.

Assignment 4: I was tasked with mining the first page of the headlines and the number of views in a url @www.nairaland. And writing the data to a csv file.